requirements.txt
reflex==0.8.24.post1
httpx
pillow
python-dotenv